style: Rename JwtService package and update imports
