+++ src/main/java/com/twoclock/gitconnect/domain/member/service/MemberAuthService.java
@@ -10,8 +10,8 @@ import com.twoclock.gitconnect.domain.member.repository.MemberRepository;
 import com.twoclock.gitconnect.global.constants.GitHubUri;
 import com.twoclock.gitconnect.global.exception.CustomException;
 import com.twoclock.gitconnect.global.exception.constants.ErrorCode;
-import com.twoclock.gitconnect.global.jwt.JwtService;
 import com.twoclock.gitconnect.global.jwt.dto.JwtTokenInfoDto;
+import com.twoclock.gitconnect.global.jwt.service.JwtService;
 import com.twoclock.gitconnect.global.security.UserDetailsImpl;
 import com.twoclock.gitconnect.global.util.RestClientUtil;
 import jakarta.servlet.http.Cookie;
